backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain-groq==0.3.7
pydantic>=2.7.4
pydantic-settings>=2.1.0
python-multipart==0.0.6
sqlalchemy==2.0.23
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
jinja2==3.1.2
